Refactor some packages
