+++ src/java/test/com/mailler/controller/sender/EmailToMimeMessageConverterTest.java
@@ -1,0 +1,5 @@
+package com.mailler.controller.sender;
+
+public class EmailToMimeMessageConverterTest {
+
+}
